1. disable integ test if access token not set over sys env 2. move config properties to be commandline options 3. add null check to stat comparators

+++ src/main/java/com/example/githuborgcli/Repository.java
@@ -30,8 +30,8 @@ public class Repository {
         contribution = (pullRequestCount > 0.0) ? pullRequestCount / forks : 0.0;
     }
 
-    public Double getStars() {
-        return (double) ghRepository.getStargazersCount();
+    public Integer getStars() {
+        return ghRepository.getStargazersCount();
     }
 
     public int getPullRequestCount() {
